+++ SettingsDialog.ts
@@ -1,0 +1,162 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1">
+<context>
+    <name>SettingsDialog</name>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="126"/>
+        <source>Settings</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="130"/>
+        <source>General</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="131"/>
+        <source>Open URL in Web Browser after playlist was generated</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="136"/>
+        <source>Copy URL to clipboard after playlist was generated</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="141"/>
+        <source>Language</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="142"/>
+        <source>Theme</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="143"/>
+        <source>Reset to defaults</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="147"/>
+        <source>Keyboard Shortcuts</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
+        <source>Command</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
+        <source>Description</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
+        <source>Keyboard shortcut</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="160"/>
+        <source>Create new playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="161"/>
+        <source>Open existing .ytplaylist file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="162"/>
+        <source>Save opened playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="163"/>
+        <source>Save playlist as .ytplaylist file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="164"/>
+        <source>Import items from .txt or .csv file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="165"/>
+        <source>Export items to .txt or .csv file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="166"/>
+        <source>Add new item (video id or URL) to playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="169"/>
+        <source>Delete selected item from playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="170"/>
+        <source>Rename selected item in playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="171"/>
+        <source>Apply shuffle mode to playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="172"/>
+        <source>Generate playlist URL</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="173"/>
+        <source>Sort items ascending</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="174"/>
+        <source>Sort items descending</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="175"/>
+        <source>Count all items in playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="176"/>
+        <source>Remove all items in playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="177"/>
+        <source>Display video information to selected video id</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="180"/>
+        <source>Remove any duplicates in playlist</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="181"/>
+        <source>Copy generated playlist URL</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="182"/>
+        <source>Search for videos by its title</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="src/main/python/dialogs/settings_dialog.py" line="193"/>
+        <source>Change</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+</TS>