+++ src/main/resources/base/forms/translations/de/SettingsDialog.ts
@@ -4,153 +4,157 @@
   <context>
     <name>SettingsDialog</name>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="126"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="126"/>
       <source>Settings</source>
       <translation>Einstellungen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="129"/>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="128"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="130"/>
       <source>General</source>
       <translation>Allgemein</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="131"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="131"/>
       <source>Open URL in Web Browser after playlist was generated</source>
       <translation>URL im Web-Browser öffnen, nachdem die Playlist generiert wurde</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="136"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="136"/>
       <source>Copy URL to clipboard after playlist was generated</source>
       <translation>URL in die Zwischenablage kopieren, nachdem die Playlist generiert wurde</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="140"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="141"/>
       <source>Language</source>
       <translation>Sprache</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="141"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="142"/>
       <source>Theme</source>
       <translation>Erscheinungsbild</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="143"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="143"/>
       <source>Reset to defaults</source>
       <translation>Auf Standardwerte zurücksetzen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="147"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="147"/>
       <source>Keyboard Shortcuts</source>
       <translation>Tastenkombinationen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="152"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Command</source>
       <translation>Kommando</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="153"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Description</source>
       <translation>Beschreibung</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="154"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Keyboard shortcut</source>
       <translation>Tastenkombination</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="159"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="160"/>
       <source>Create new playlist</source>
       <translation>Neue Playlist erstellen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="160"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="161"/>
       <source>Open existing .ytplaylist file</source>
       <translation>Vorhandene .ytplaylist-Datei öffnen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="161"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="162"/>
       <source>Save opened playlist</source>
       <translation>Aktuelle Playlist speichern</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="162"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="163"/>
       <source>Save playlist as .ytplaylist file</source>
       <translation>Playlist als .ytplaylist-Datei speichern</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="163"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="164"/>
       <source>Import items from .txt or .csv file</source>
       <translation>Einträge aus .txt oder .csv Datei importieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="164"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="165"/>
       <source>Export items to .txt or .csv file</source>
       <translation>Einträge in .txt oder .csv Datei exportieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="165"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="166"/>
       <source>Add new item (video id or URL) to playlist</source>
       <translation>Neuen Eintrag (Video-ID oder URL) zur Playlist hinzufügen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="168"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="169"/>
       <source>Delete selected item from playlist</source>
       <translation>Ausgewählten Eintrag aus Playlist löschen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="169"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="170"/>
       <source>Rename selected item in playlist</source>
       <translation>Ausgewählten Eintrag in Playlist umbenennen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="170"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="171"/>
       <source>Apply shuffle mode to playlist</source>
       <translation>Shuffle-Modus auf Playlist anwenden</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="171"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="172"/>
       <source>Generate playlist URL</source>
       <translation>Playlist-URL generieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="172"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="173"/>
       <source>Sort items ascending</source>
       <translation>Einträge aufsteigend sortieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="173"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="174"/>
       <source>Sort items descending</source>
       <translation>Einträge absteigend sortieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="174"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="175"/>
       <source>Count all items in playlist</source>
       <translation>Alle Einträge in Playlist zählen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="175"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="176"/>
       <source>Remove all items in playlist</source>
       <translation>Alle Einträge aus Playlist entfernen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="176"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="177"/>
       <source>Display video information to selected video id</source>
       <translation>Videoinformationen zur ausgewählten Video-ID anzeigen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="179"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="180"/>
       <source>Remove any duplicates in playlist</source>
       <translation>Duplikate in der Playlist entfernen</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="180"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="181"/>
       <source>Copy generated playlist URL</source>
       <translation>Generierte Playlist-URL kopieren</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="191"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="182"/>
+      <source>Search for videos by its title</source>
+      <translation>Suche nach Videos nach dem Titel</translation>
+    </message>
+    <message>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="193"/>
       <source>Change</source>
       <translation>Ändern</translation>
     </message>