+++ src/main/resources/base/forms/translations/es-ES/SettingsDialog.ts
@@ -4,153 +4,157 @@
   <context>
     <name>SettingsDialog</name>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="126"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="126"/>
       <source>Settings</source>
       <translation>Configuración</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="129"/>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="128"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="130"/>
       <source>General</source>
       <translation>General</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="131"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="131"/>
       <source>Open URL in Web Browser after playlist was generated</source>
       <translation>Abrir URL en el navegador web después de generar la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="136"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="136"/>
       <source>Copy URL to clipboard after playlist was generated</source>
       <translation>Copiar URL al portapapeles después de generar la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="140"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="141"/>
       <source>Language</source>
       <translation>Idioma</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="141"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="142"/>
       <source>Theme</source>
       <translation>Apariencia</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="143"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="143"/>
       <source>Reset to defaults</source>
       <translation>Restablecer valores predeterminados</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="147"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="147"/>
       <source>Keyboard Shortcuts</source>
       <translation>Atajos de teclado</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="152"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Command</source>
       <translation>Comando</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="153"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Description</source>
       <translation>Descripción</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="154"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="151"/>
       <source>Keyboard shortcut</source>
       <translation>Atajos de teclado</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="159"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="160"/>
       <source>Create new playlist</source>
       <translation>Crear nueva playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="160"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="161"/>
       <source>Open existing .ytplaylist file</source>
       <translation>Abrir archivo .ytplaylist existente</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="161"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="162"/>
       <source>Save opened playlist</source>
       <translation>Guardar la playlist actual</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="162"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="163"/>
       <source>Save playlist as .ytplaylist file</source>
       <translation>Guardar playlist como archivo .ytplaylist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="163"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="164"/>
       <source>Import items from .txt or .csv file</source>
       <translation>Importar ítems de archivo .txt o .csv</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="164"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="165"/>
       <source>Export items to .txt or .csv file</source>
       <translation>Exportar ítems a archivo .txt o .csv</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="165"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="166"/>
       <source>Add new item (video id or URL) to playlist</source>
       <translation>Añadir un nuevo ítem (id de vídeo o URL) a la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="168"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="169"/>
       <source>Delete selected item from playlist</source>
       <translation>Borrar el ítem seleccionado de la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="169"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="170"/>
       <source>Rename selected item in playlist</source>
       <translation>Renombrar ítem seleccionado en la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="170"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="171"/>
       <source>Apply shuffle mode to playlist</source>
       <translation>Aplicar shuffle a la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="171"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="172"/>
       <source>Generate playlist URL</source>
       <translation>Generar URL de playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="172"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="173"/>
       <source>Sort items ascending</source>
       <translation>Ordenar ítems ascendentes</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="173"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="174"/>
       <source>Sort items descending</source>
       <translation>Ordenar ítems descendentes</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="174"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="175"/>
       <source>Count all items in playlist</source>
       <translation>Contar todos los ítems de la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="175"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="176"/>
       <source>Remove all items in playlist</source>
       <translation>Borrar todos los ítems de la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="176"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="177"/>
       <source>Display video information to selected video id</source>
       <translation>Ver información de vídeo a la id de vídeo seleccionado</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="179"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="180"/>
       <source>Remove any duplicates in playlist</source>
       <translation>Eliminar duplicados en la playlist</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="180"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="181"/>
       <source>Copy generated playlist URL</source>
       <translation>Copiar URL de la playlist generada</translation>
     </message>
     <message>
-      <location filename="../../../../python/dialogs/settings_dialog.py" line="191"/>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="182"/>
+      <source>Search for videos by its title</source>
+      <translation>Buscar videos por título</translation>
+    </message>
+    <message>
+      <location filename="src/main/python/dialogs/settings_dialog.py" line="193"/>
       <source>Change</source>
       <translation>Cambiar</translation>
     </message>