+++ src/main/resources/base/forms/translations/nl/SearchDialog.ts
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="nl" sourcelanguage="en">
+  <context>
+    <name>SearchDialog</name>
+    <message>
+      <location filename="src/main/python/dialogs/search_dialog.py" line="291"/>
+      <source>Search</source>
+      <translation>Zoeken</translation>
+    </message>
+  </context>
+</TS>