+++ src/main/resources/base/forms/translations/pl/VideoInfoDialog.ts
@@ -4,7 +4,7 @@
   <context>
     <name>VideoInfoDialog</name>
     <message>
-      <location filename="../../../../python/dialogs/video_info_dialog.py" line="102"/>
+      <location filename="src/main/python/dialogs/video_info_dialog.py" line="109"/>
       <source>Video information</source>
       <translation>Informacje o filmiku</translation>
     </message>